FakeReviewDetection_App/requirements.txt
streamlit
torch
transformers
scikit-learn
joblib
numpy
pandas
matplotlib
seaborn
pillow
